+++ libft/ft_hexlen.c
@@ -6,20 +6,19 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/17 20:59:17 by mfeldman          #+#    #+#             */
-/*   Updated: 2022/07/20 21:57:01 by mfeldman         ###   ########.fr       */
+/*   Updated: 2022/07/21 01:21:38 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_hexlen(unsigned int nb)
+size_t	ft_hexlen(unsigned long long int nb)
 {
 	size_t	i;
 
 	i = 0;
 	if (nb == 0)
 		return(1);
-	
 	while (nb)
 	{
 		i++;
